demo_requirements.txt
django
django-js-asset
selenium
tox
